Clean up reference designs and documentation

+++ wp-content/plugins/paid-memberships-pro/adminpages/membershiplevels.php
@@ -176,18 +176,21 @@
 		}
 
 		// For each group, make sure that each level in the group still exists. If not, remove the link.
-		foreach ( $level_groups as $level_group ) {
-			$group_level_ids = pmpro_get_level_ids_for_group( $level_group->id );
-			foreach ( $group_level_ids as $group_level_id ) {
-				$level_exists = false;
-				foreach ( $reordered_levels as $reordered_level ) {
-					if ( $group_level_id === $reordered_level->id ) {
-						$level_exists = true;
-						break;
+		// We only want to do this if we are not searching for specific levels, otherwise $reordered_levels may not contain all levels.
+		if ( empty( $s) ) {
+			foreach ( $level_groups as $level_group ) {
+				$group_level_ids = pmpro_get_level_ids_for_group( $level_group->id );
+				foreach ( $group_level_ids as $group_level_id ) {
+					$level_exists = false;
+					foreach ( $reordered_levels as $reordered_level ) {
+						if ( $group_level_id === $reordered_level->id ) {
+							$level_exists = true;
+							break;
+						}
 					}
-				}
-				if ( ! $level_exists ) {
-					$wpdb->query( $wpdb->prepare( "DELETE FROM $wpdb->pmpro_membership_levels_groups WHERE `group` = %d AND `level` = %d", $level_group->id, $group_level_id ) );
+					if ( ! $level_exists ) {
+						$wpdb->query( $wpdb->prepare( "DELETE FROM $wpdb->pmpro_membership_levels_groups WHERE `group` = %d AND `level` = %d", $level_group->id, $group_level_id ) );
+					}
 				}
 			}
 		}
@@ -289,7 +292,13 @@
 			</form>
 
 			<?php if(empty($_REQUEST['s']) && count($reordered_levels) > 1) { ?>
-				<p><?php esc_html_e('Drag and drop membership levels within the group to reorder them on the Membership Levels page. Reorder groups using the up/down arrows.', 'paid-memberships-pro' ); ?></p>
+				<p><?php
+					esc_html_e('Drag and drop membership levels within the group to reorder them on the Membership Levels page. Reorder groups using the up/down arrows.', 'paid-memberships-pro' );
+					echo ' ';
+					$membership_levels_link = '<a title="' . esc_attr__( 'Paid Memberships Pro - Membership Levels Documentation', 'paid-memberships-pro' ) . '" target="_blank" rel="nofollow noopener" href="https://www.paidmembershipspro.com/documentation/admin/levels-settings/?utm_source=plugin&utm_medium=pmpro-membershiplevels&utm_campaign=documentation&utm_content=&utm_term=">' . esc_html__( 'Membership Levels', 'paid-memberships-pro' ) . '</a>';
+					// translators: %s: Link to Membership Levels doc.
+					printf( esc_html__('Learn more about %s.', 'paid-memberships-pro' ), $membership_levels_link ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+				?></p>
 			<?php } ?>
 
 			<?php
@@ -394,7 +403,7 @@
 												$delete_text = esc_html(
 													sprintf(
 														// translators: %s is the Level Name.
-														__( "Are you sure you want to delete membership level %s? Any gateway subscriptions or third-party connections with a member's account will remain active.", 'paid-memberships-pro' ),
+														__( "Are you sure you want to delete membership level %s? All payment subscriptions for this level will be cancelled.", 'paid-memberships-pro' ),
 														$level->name
 													)
 												);
